separated into ListConnectors and ListOperations
